requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.3
Werkzeug==2.3.0
gunicorn==20.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
